+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 ## Changelog
 
+### Version 1.0.1
+- **Feature**: New functionality to assign attribute set by category.
+- **Feature**: New functionality to assign product to category by attribute set.
+- **Feature**: New functionality to assign product to category by category ID.
+- **Feature**: New functionality to clean-up product super attributes.
+
 ### Version 1.0.0
 - **Feature**: New module to facilitate general utilities.